Creation of exercises is operational. Adding an exercise to a workout is operational.
Next, after an exercise is added, the user needs to be taken back to the editexercises screen.
After that, They need to be able to remove exercises from that listview.
Also, selection from a listview (taking the object and moving to a different screen) needs to happen at some point.